+++ stu/main.go
@@ -5,32 +5,32 @@ import (
 	"ny/dorm"
 )
 
-func GetStuDormIdByUserId(userId int) (bool, int) {
+func GetStuDormIdByUserId(userId int64) (bool, int64) {
 	var s stu
 	s.UserId = userId
 	return s.queryByUserId(), s.DormId
 }
 
-func GetUserIdByStuId(stuId int) (bool, int) {
+func GetUserIdByStuId(stuId int64) (bool, int64) {
 	var s stu
 	s.Id = stuId
 	r := s.queryById()
 	return r, s.UserId
 }
 
-func GetStuDormIdById(id int) (bool, int) {
+func GetStuDormIdById(id int64) (bool, int64) {
 	var s stu
 	s.Id = id
 	return s.queryById(), s.DormId
 }
 
-func GetStuNumber(userId int) (bool, string) {
+func GetStuNumber(userId int64) (bool, string) {
 	var s stu
 	s.UserId = userId
 	return s.queryByUserId(), s.StuNumber
 }
 
-func SaveStuNumber(userId int, stuNumber string) bool {
+func SaveStuNumber(userId int64, stuNumber string) bool {
 	var s stu
 	s.UserId = userId
 	s.queryByUserId()
@@ -38,7 +38,7 @@ func SaveStuNumber(userId int, stuNumber string) bool {
 	return s.update()
 }
 
-func SaveDorm(userId int, dormId int) bool {
+func SaveDorm(userId int64, dormId int64) bool {
 	var s stu
 	s.UserId = userId
 	r := s.queryByUserId()
@@ -46,7 +46,7 @@ func SaveDorm(userId int, dormId int) bool {
 	return r && s.update()
 }
 
-func SaveRoom(userId int, room string) bool {
+func SaveRoom(userId int64, room string) bool {
 	var s stu
 	s.UserId = userId
 	r := s.queryByUserId()
@@ -54,7 +54,7 @@ func SaveRoom(userId int, room string) bool {
 	return r && s.update()
 }
 
-func AddStu(userId int) (bool, int) {
+func AddStu(userId int64) (bool, int64) {
 	var s stu
 	var ok1 bool
 	s.UserId = userId
@@ -62,14 +62,14 @@ func AddStu(userId int) (bool, int) {
 	return ok1 && s.insert(), s.Id
 }
 
-func GetStuExitsByUserId(userId int) bool {
+func GetStuExitsByUserId(userId int64) bool {
 	var s stu
 	s.UserId = userId
 	r := s.queryByUserId()
 	return r
 }
 
-func GetStuIdByUserId(userId int) (bool, int) {
+func GetStuIdByUserId(userId int64) (bool, int64) {
 	var s stu
 	s.UserId = userId
 
@@ -77,13 +77,13 @@ func GetStuIdByUserId(userId int) (bool, int) {
 
 }
 
-func GetStuExitsById(id int) bool {
+func GetStuExitsById(id int64) bool {
 	var s stu
 	s.Id = id
 	return s.queryById()
 }
 
-func GetStuRoomByUserId(id int) (bool, string) {
+func GetStuRoomByUserId(id int64) (bool, string) {
 	var s stu
 	s.UserId = id
 	r := s.queryByUserId()
